fix: address PRRT_kwDOSJAM6s6HqDlR — split pipeline/cache-key helpers into parsing module to stay under line limit
Move the module-level pipeline-matching and cache-key helpers
(_as_dict, is_pipeline_owned_status, pipeline_targets_branch,
pipeline_has_ref_info, pipeline_targets_pr, freeze_params) and the
_FrozenParams alias out of bitbucket_client.py into the existing
bitbucket_client_parsing.py. This brings bitbucket_client.py from 1514
to 1451 lines, under MAX_FIRST_PARTY_FILE_LINES (1500) enforced by
test_first_party_code_files_stay_under_line_limit.
